refactor: use `store/persistence`
