+++ packages/playground/examples/blob/index.html
@@ -33,7 +33,7 @@
     </div>
     <br />
     <script type="module">
-      import '@blocksuite/store/src/blob/__tests__/test-entry';
+      import '@blocksuite/store/src/persistence/blob/__tests__/test-entry';
     </script>
   </body>
 </html>
