+++ packages/store/src/index.ts
@@ -3,10 +3,10 @@ export * from './space.js';
 export * from './store.js';
 export * from './base.js';
 export * from './awareness.js';
-export * from './blob/index.js';
+export * from './persistence/blob/index.js';
 export * from './text-adapter.js';
 export * from '@blocksuite/global/utils';
-export * from './doc-providers.js';
+export * from './persistence/doc/index.js';
 export * from './workspace/index.js';
 export * as Utils from './utils/utils.js';
 export {
