+++ packages/store/src/persistence/blob/__tests__/blob.spec.ts
@@ -1,5 +1,5 @@
 import { test } from '@playwright/test';
-import { collectTestResult } from '../../__tests__/test-utils-node.js';
+import { collectTestResult } from '../../../__tests__/test-utils-node.js';
 
 // checkout test-entry.ts for actual test cases
 const blobExamplePage = 'http://localhost:5173/examples/blob/';
