+++ packages/store/src/persistence/blob/utils.ts
